% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataValuestoLocal.R
\name{addDataValuestoLocal}
\alias{addDataValuestoLocal}
\title{addDataValuestoLocal}
\usage{
addDataValuestoLocal(con, SiteCode, VariableCode, update = FALSE)
}
\arguments{
\item{con}{DB connection}

\item{SiteCode}{}

\item{update}{boolean comportamiento ante registro duplicado (si TRUE, actualiza)}

\item{TimeSeries}{data.frame resultante de GetValues (WaterML Package)}
}
\value{
"Number rows inserted or updated:..." or "nothing inserted/updated"
}
\description{
Esta funcion inserta en ODM una serie temporal contenida en un dataframe como el producido por GetValues del package WaterML.
}
