% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilbambu.R
\name{addSeriestoLocal}
\alias{addSeriestoLocal}
\title{addSeriestoLocal}
\usage{
addSeriestoLocal(params, update = FALSE)
}
\arguments{
\item{params}{lista de parametros (fila del data.frame)}

\item{update}{boolean comportamiento ante VariableCode+SiteCode duplicado (si TRUE, actualiza)}
}
\value{
""1 row inserted or updated. SeriesID:..." or "nothing inserted/updated"
}
\description{
Esta funcion inserta en ODM una Serie contenida en un dataframe como el producido por GetSiteInfo del package WaterML.
}
