% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSitestoLocal.R
\name{addSitestoLocal}
\alias{addSitestoLocal}
\title{addSitestoLocal}
\usage{
addSitestoLocal(con, sites, update = FALSE)
}
\arguments{
\item{sites}{data.frame resultante de GetSites (WaterML)}

\item{update}{boolean comportamiento ante VariableName duplicado (si TRUE, actualiza)}
}
\value{
"N rows inserted or updated. VariableID:..." or "nothing inserted/updated"
}
\description{
Esta funcion inserta en ODM un Site contenido en un dataframe como el producido por GetSites del package WaterML.
}
