% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilbambu.R
\name{addVariabletoLocal}
\alias{addVariabletoLocal}
\title{addVariabletoLocal}
\usage{
addVariabletoLocal(params, update = FALSE)
}
\arguments{
\item{params}{lista de parametros (fila del data.frame)}

\item{update}{boolean comportamiento ante VariableName duplicado (si TRUE, actualiza)}
}
\value{
1 row inserted or updated. VariableID:..." or "nothing inserted/updated"
}
\description{
Esta funcion inserta en ODM una variable contenida en un dataframe como el producido por GetVariables del package WaterML.
}
