% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSiteInfoFromODM.R
\name{GetSiteInfoFromODM}
\alias{GetSiteInfoFromODM}
\title{GetSiteInfoFromODM}
\usage{
GetSiteInfoFromODM(con, SiteCode, VariableCode = NULL, MethodID = NULL,
  SourceID = NULL)
}
\arguments{
\item{con}{Conexión a base de datos (creada con dbConnect)}

\item{SiteCode}{código de sitio}

\item{VariableCode}{código de variable (optativo)}

\item{MethodID}{ID de método (optativo)}

\item{SourceID}{ID de fuente (optativo)}
}
\value{
un data.frame con las siguientes columnas: SiteID SiteName SiteCode FullSiteCode Latitude 
 Longitude Elevation_m State County Country Comments FeatureType VariableCode FullVariableCode  VariableName
ValueType  DataType GeneralCategory SampleMedium UnitName UnitType UnitAbbreviation NoDataValue
 IsRegular TimeUnitName TimeUnitAbbreviation TimeSupport
Speciation methodID methodCode methodDescription methodLink sourceID organization sourceDescription 
citation qualityControlLevelID 
qualityControlLevelCode qualityControlLevelDefinition valueCount beginDateTime
 endDateTime beginDateTimeUTC endDateTimeUTC
}
\description{
Esta función extrae la tabla de variables observadas para un sitio seleccionado de una base de datos ODM
Se debe indicar el SiteCode y permite filtrar por VariableCode, MethodID y SourceID.
Devuelve un data.frame con las filas como definiciones de series y columnas como sus propiedades.
El mismo sirve como parámetro de entrada para la función GetValuesFromODM
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='11.22.33.444',dbname='ODM')
GetSiteInfoFromODM(con,VariableCode='2',MethodID=1)
}
\keyword{GetSiteInfo}
\keyword{ODM}
\keyword{series}
\keyword{seriesCatalog}
