% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSitesFromODM.R
\name{GetSitesFromODM}
\alias{GetSitesFromODM}
\title{GetSitesFromODM}
\usage{
GetSitesFromODM(con, SiteCode = NULL, FeatureType = NULL, north = NULL,
  south = NULL, east = NULL, west = NULL)
}
\arguments{
\item{con}{Conexión a base de datos (creada con dbConnect)}

\item{SiteCode}{Código del sitio (uno o mas 'a:int' o 'p:int', opcional)}

\item{FeatureType}{tipo de objeto espacial ('point' o 'area', 'grid', 'line', opcional)}

\item{north}{coordenada norte del recuadro (opcional)}

\item{south}{coordenada sur del recuadro (opcional)}

\item{east}{coordenada este del recuadro (opcional)}

\item{west}{coordenada oeste del recuadro (opcional)}
}
\value{
un data.frame con las siguientes columnas:  SiteID, SiteCode, SiteName, Latitude, Longitude, Elevation_m,
State, County, Comments, FeatureType
}
\description{
Esta función extrae la tabla Sites de la base de datos ODM. Devuelve un data.frame cuyas filas corresponden a los sitios y las
columnas a las propiedades, el cual sirve como parámetro de entrada para la función extractSeriesCatalog.
 Se puede obtener el listado completo o filtrar por SiteCode, FeatureType o recuadro espacial (north,south,
 east,west)
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='11.22.33.444',dbname='ODM')
sites<-GetSitesFromODM(con,featureType='point',north=-20,south=-25,east=-55,west=-60)
}
\keyword{GetSites}
\keyword{ODM}
\keyword{Sites}
