% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetValuesFromODM.R
\name{GetValuesFromODM}
\alias{GetValuesFromODM}
\title{GetValuesFromLocal}
\usage{
GetValuesFromODM(con, SeriesID = NULL, SiteCode = NULL,
  VariableCode = NULL, StartDate, EndDate)
}
\arguments{
\item{con}{conexión con la base de datos (creada mediante dbConnect)}

\item{SeriesID}{integer o vector de integer (obtenidos de GetSiteInfoFromLocal)}

\item{SiteCode}{string o vector de string (solo si SeriesID es NULL)}

\item{VariableCode}{string o vector de string (solo si SeriesID es NULL is SiteCode no es NULL)}

\item{StartDate}{ISO datetime UTC}

\item{EndDate}{ISO datetime UTC}
}
\value{
data.frame con las siguientes columnas: DateTime DataValue UTCOffset CensorCode DateTimeUTC MethodID SourceID QualityControlLevelCode SeriesID SiteCode VariableCode VariableName
}
\description{
Esta función descarga los pares de fecha y valor de una base de datos ODM correspondientes a las series indicadas
usando SeriesID o VariableCode+SiteCode, entre las fechas StartDate y EndDate.
Devuelve un data.frame
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='11.22.33.444',dbname='ODM')
SiteInfo<-GetSiteInfoFromODM(con,SiteCode='alturas_dinac:155',VariableCode='2')
GetValuesFromODM(con,SiteInfo,'2010-01-01','2015-01-01')
}
\keyword{GetValues}
\keyword{ODM}
\keyword{TimeSeries}
\keyword{values}
