% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestSeriesCatalog.R
\name{harvestSeriesCatalog}
\alias{harvestSeriesCatalog}
\title{harvestSeriesCatalog}
\usage{
harvestSeriesCatalog(url, SiteCode, update = FALSE)
}
\arguments{
\item{url}{WOFWML WS end point}

\item{SiteCode}{FullSiteCode from GetSites}

\item{update}{boolean comportamiento ante SiteID*VariableID duplicado (si TRUE, actualiza)}
}
\value{
""1 row inserted or updated. SeriesID:..." or "nothing inserted/updated"
}
\description{
Esta funcion descarga SeriesCatalog de un WOFWML WS e inserta en ODM usando "addSeriestoLocal"
}
\examples{
harvestSeriesCatalog("http://brasilia.essi-lab.eu/gi-axe/services/cuahsi_1_1.asmx?WSDL",TRUE)
}
