% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilbambu.R
\name{harvestSites}
\alias{harvestSites}
\title{harvestSites}
\usage{
harvestSites(url, north = 90, south = -90, east = 180, west = -180,
  update = FALSE)
}
\arguments{
\item{url}{WOFWML WS end point}

\item{update}{boolean comportamiento ante SiteID duplicado (si TRUE, actualiza)}
}
\value{
""1 row inserted or updated. SiteID:..." or "nothing inserted/updated"
}
\description{
Esta funcion descarga Sites de un WOFWML WS e inserta en ODM usando "addSitesToLocal"
}
