% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestVariables.R
\name{harvestVariables}
\alias{harvestVariables}
\title{harvestVariables}
\usage{
harvestVariables(url, update = FALSE)
}
\arguments{
\item{url}{WOFWML WS end point}

\item{update}{boolean comportamiento ante VariableID duplicado (si TRUE, actualiza)}
}
\value{
"1 row inserted or updated. VariableID:..." or "nothing inserted/updated"
}
\description{
Esta funcion descarga Variables de un WOFWML WS e inserta en ODM usando "addVariableToLocal"
}
\examples{
harvestVariables("http://brasilia.essi-lab.eu/gi-axe/services/cuahsi_1_1.asmx?WSDL",TRUE)
}
