% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotValuesODM.R
\name{PlotValuesODM}
\alias{PlotValuesODM}
\title{PlotValuesODM}
\usage{
PlotValuesODM(Values, output = NULL, types = c("l"), width = 800,
  height = 700, invert = c(FALSE), StartDate = NULL, EndDate = NULL)
}
\arguments{
\item{Values}{data.frame obtenido mediante getValues (pueden concatenarse dos o mas)}

\item{output}{Archivo donde se imprima el gráfico (opcional). Por defecto abre un display x11}

\item{types}{vector Tipos de gráfico para cada variable ('l': línea, 'p': puntos, 's' escalones), en el orden de aparición (opcional)}

\item{width}{Ancho del gráfico (opcional)}

\item{height}{Altura del gráfico (opcional)}

\item{invert}{vector de boolean Determina si debe invertirse el eje Y para cada variable (opcional)}

\item{StartDate}{fecha inicial (opcional)}

\item{EndDate}{fecha final (opcional)}
}
\value{
0
}
\description{
Esta función grafica las series temporales obtenidas mediante GetValuesFromODM
Admite hasta dos variables (variableName) distintas, sin límite para el número
de se sitios de cada variable.
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='11.22.33.444',dbname='ODM')
Values1<-GetValuesFromODM(con,SiteCode='alturas_dinac:155',VariableCode='2',StartDate='2010-01-01',EndDate='2015-01-01')
Values2<-GetValuesFromODM(con,SiteCode='alturas_dinac:153',VariableCode='2',StartDate='2010-01-01',EndDate='2015-01-01')
Values<-rbind(Values1,Values2) 
PlotValuesODM(values,seriesCatalog,output='~/tmp/plotValues.png')
}
\keyword{ODM}
\keyword{plot}
\keyword{timeSeries}
\keyword{values}
